legacy: migrate aligned memory file reader to Go
Replace the C implementation of nox_memfile_read64align_40AD60 with an exported Go wrapper around binfile.MemFile.ReadU64Align. Preserve alignment padding, eight-byte consumption, short-read no-write behavior, and prefix copying already covered by the memory-file tests.

Declare the bridge return, size, and count with uint32_t and int32_t, and use C.uint32_t and C.int32_t instead of platform-sized integer aliases.

+++ src/legacy/noxstring_test.go
@@ -410,7 +410,7 @@ func TestNoxMemfile(t *testing.T) {
 
 	// 7. Read block
 	var buf [4]byte
-	n := uint(NoxMemfileRead(unsafe.Pointer(&buf[0]), 1, 4, cFile))
+	n := NoxMemfileRead(unsafe.Pointer(&buf[0]), 1, 4, cFile)
 	if n != 4 {
 		t.Errorf("expected read 4, got %d", n)
 	}
